+++ src/components/ui/Badge.tsx
@@ -17,11 +17,10 @@ function Badge({ className = '', variant = 'default', ...props }: BadgeProps) {
 
 	return (
 		<div
-			className={`inline-flex items-center rounded-full border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:outline-none focus:ring-2 focus:ring-gray-950 focus:ring-offset-2 dark:focus:ring-gray-300 ${variants[variant]} ${className}`}
+			className={`inline-flex items-center rounded-full border px-2.5 py-0.5 text-xs font-semibold transition-colors focus:ring-2 focus:ring-gray-950 focus:ring-offset-2 focus:outline-none dark:focus:ring-gray-300 ${variants[variant]} ${className}`}
 			{...props}
 		/>
 	)
 }
 
 export { Badge }
-
