+++ src/components/ui/Tabs.tsx
@@ -24,8 +24,7 @@ export function Tabs({
 	children,
 	className = '',
 }: TabsProps) {
-	const [uncontrolledValue, setUncontrolledValue] =
-		React.useState(defaultValue)
+	const [uncontrolledValue, setUncontrolledValue] = React.useState(defaultValue)
 	const value = controlledValue ?? uncontrolledValue
 	const handleValueChange = onValueChange ?? setUncontrolledValue
 
@@ -36,10 +35,7 @@ export function Tabs({
 	)
 }
 
-export function TabsList({
-	className = '',
-	...props
-}: React.HTMLAttributes<HTMLDivElement>) {
+export function TabsList({ className = '', ...props }: React.HTMLAttributes<HTMLDivElement>) {
 	return (
 		<div
 			className={`inline-flex h-10 items-center justify-center rounded-md bg-gray-100 p-1 text-gray-500 dark:bg-gray-800 dark:text-gray-400 ${className}`}
@@ -48,16 +44,11 @@ export function TabsList({
 	)
 }
 
-interface TabsTriggerProps
-	extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+interface TabsTriggerProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
 	value: string
 }
 
-export function TabsTrigger({
-	value,
-	className = '',
-	...props
-}: TabsTriggerProps) {
+export function TabsTrigger({ value, className = '', ...props }: TabsTriggerProps) {
 	const context = React.useContext(TabsContext)
 	if (!context) throw new Error('TabsTrigger must be used within Tabs')
 
@@ -65,10 +56,8 @@ export function TabsTrigger({
 
 	return (
 		<button
-			className={`inline-flex items-center justify-center whitespace-nowrap rounded-sm px-3 py-1.5 text-sm font-medium ring-offset-white transition-all focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-gray-950 focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 dark:ring-offset-gray-950 dark:focus-visible:ring-gray-300 ${
-				isActive
-					? 'bg-white text-gray-950 shadow-sm dark:bg-gray-950 dark:text-gray-50'
-					: ''
+			className={`inline-flex items-center justify-center rounded-sm px-3 py-1.5 text-sm font-medium whitespace-nowrap ring-offset-white transition-all focus-visible:ring-2 focus-visible:ring-gray-950 focus-visible:ring-offset-2 focus-visible:outline-none disabled:pointer-events-none disabled:opacity-50 dark:ring-offset-gray-950 dark:focus-visible:ring-gray-300 ${
+				isActive ? 'bg-white text-gray-950 shadow-sm dark:bg-gray-950 dark:text-gray-50' : ''
 			} ${className}`}
 			onClick={() => context.onValueChange(value)}
 			{...props}
@@ -80,11 +69,7 @@ interface TabsContentProps extends React.HTMLAttributes<HTMLDivElement> {
 	value: string
 }
 
-export function TabsContent({
-	value,
-	className = '',
-	...props
-}: TabsContentProps) {
+export function TabsContent({ value, className = '', ...props }: TabsContentProps) {
 	const context = React.useContext(TabsContext)
 	if (!context) throw new Error('TabsContent must be used within Tabs')
 
@@ -92,9 +77,8 @@ export function TabsContent({
 
 	return (
 		<div
-			className={`mt-2 ring-offset-white focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-gray-950 focus-visible:ring-offset-2 dark:ring-offset-gray-950 dark:focus-visible:ring-gray-300 ${className}`}
+			className={`mt-2 ring-offset-white focus-visible:ring-2 focus-visible:ring-gray-950 focus-visible:ring-offset-2 focus-visible:outline-none dark:ring-offset-gray-950 dark:focus-visible:ring-gray-300 ${className}`}
 			{...props}
 		/>
 	)
 }
-
